+++ xxw-shop-applications/xxw-shop-service/xxw-shop-service-rbac/src/main/java/com/xxw/shop/module/menu/entity/MenuPermission.java
@@ -27,7 +27,7 @@ public class MenuPermission implements Serializable {
      * 菜单资源用户id
      */
     @Id(keyType = KeyType.Auto)
-    private BigInteger menuPermissionId;
+    private Long menuPermissionId;
     /**
      * 创建时间
      */
